Updates itsre-deploy module to be compatible with terraform 0.12
This will be a huge change but its mostly related to updating the
codebase to support terraform 0.12. Most of the changes were generated
automatically using the `terraform 0.12upgrade -yes` command which
converts the code for you

There are a couple of changes in here that are a refactor of feature
that was not available in terraform 0.11 and is not available in
terraform 0.12. See below:

 - Removed the `vpc/subnet` module, we no longer need to calculate
 subnets that way since we can do for loops now which simplifies things
 - Upgrade upstream vpc module from terraform registry to latest
 version, this newer version supports version 0.12
 - Fixed output for private and public subnets, terraform 0.12 now
 supports proper typing now so we don't need to explicitly state that
 its a list

Misc changes:

 - Remove gozer's username as a default IAM user to create
 - Default to not create dns zones when an account gets deployed.